Move stuff around to hopefully make relative imports easier. Add dnspython to requirements
