+++ http/web/py/requirements.txt
@@ -1,3 +1,4 @@
 Flask==1.1.2
 Flask-Cors==3.0.8
 pymongo==3.10.1
+dnspython==1.16.0
